refatora endpoints
