+++ src/notas/matriculas.csv
@@ -1,0 +1,11 @@
+aluno_identificador,aluno_nome
+1001,Ana Silva
+1002,Bruno Costa
+1003,Camila Souza
+1004,Daniel Pereira
+1005,Elisa Fernandes
+1006,Felipe Rocha
+1007,Gabriela Lima
+1008,Hugo Martins
+1009,Isabela Torres
+1010,João Carvalho
